+++ kh13fb/src/main/java/com/kh/kh13fb/dao/AttachDao.java
@@ -14,8 +14,7 @@ public class AttachDao {
 		@Autowired
 		private SqlSession sqlSession;
 		
-		@Autowired
-		private AttachDto attachDto;
+		
 
 		public int getSequence() {
 			return sqlSession.selectOne("attach.sequence");
